Project description:
The objective of the project is to calculate the duration of how long a person lived based on his Age.
The duration should be calculated in different time units like Months, Weeks, Days, Hours, Minutes, Seconds.
The input to the application is an Age of a person and the output is expected to calculate the duration of time he lived in different time units as selected by the user.


Below is the expected output.
What's your age? 10
Please choose time unit: Months, Weeks, Days, Hours, Minutes, Seconds. 
Note: You can write the first letter or the full name of the time unit. m
You lived for 120 Months